fix(story-4-5): code review fixes - Price float32→int64, CouponListByCartData type safety, CartModel key fix, couponSubtotal int64, nil guard
- H1: CarItemtPromotionListData.Price & CartPromotionItemList.Price float32→int64
- H4: CouponListByCartData EnableList/DisableList interface{}→[]CouponData
- H4: OrderDetailModel.CouponHistoryDetailList interface{}→CouponListByCartData
- H5: Flutter CartModel map key productId→cart item id
- M1: couponSubtotal/cartItemAmount float64→int64
- M2: order_submit buildCoupon points hardcode→dynamic
- L1: getSkuStock nil guard in handleNoReduce
- H2/H3: TODO comments for missing proto fields (LockStock/GiftPoint/GiftGrowth)
- Flutter models: CartPromotionData.price & ConfirmCartItem.price double→int
- Story status: Ready for Dev→In Progress, added code review record

+++ api/front/internal/logic/order/order/generate_confirm_order_logic.go
@@ -2,6 +2,7 @@ package order
 
 import (
 	"context"
+
 	"github.com/feihua/zero-admin/api/front/internal/logic/common"
 	"github.com/feihua/zero-admin/api/front/internal/logic/member/coupon"
 	"github.com/feihua/zero-admin/api/front/internal/logic/order/cart"
@@ -115,14 +116,14 @@ func (l *GenerateConfirmOrderLogic) GenerateConfirmOrder(req *types.GenerateConf
 	settingInfo, _ := l.svcCtx.MemberConsumeSettingService.QueryMemberConsumeSettingDetail(l.ctx, &umsclient.QueryMemberConsumeSettingDetailReq{
 		Id: 1,
 	})
-	// 6.计算总金额、活动优惠、应付金额
-	var totalAmount float32 = 0
-	var freightAmount float32 = 0
-	var promotionAmount float32 = 0
-	var payAmount float32 = 0
+	// 6.计算总金额、活动优惠、应付金额（统一使用 int64，单位：分）
+	var totalAmount int64 = 0
+	var freightAmount int64 = 0
+	var promotionAmount int64 = 0
+	var payAmount int64 = 0
 	for _, item := range cartPromotionItemList {
-		totalAmount = totalAmount + item.Price*float32(item.Quantity)
-		promotionAmount = promotionAmount + float32(item.ReduceAmount)*float32(item.Quantity)
+		totalAmount = totalAmount + item.Price*int64(item.Quantity)
+		promotionAmount = promotionAmount + item.ReduceAmount*int64(item.Quantity)
 	}
 	payAmount = totalAmount - promotionAmount
 	return &types.GenerateConfirmOrderResp{
